+++ Isolator.Fody.ConsoleApp.Tests/ClassTests.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics;
 using Isolator.ConsoleApp;
 
-namespace Isolator.Fody.Tests
+namespace Isolator.Fody.ConsoleApp.Tests
 {
     public class ClassTests
     {